+++ client/public/pageStyles.css
@@ -1,0 +1,152 @@
+/* Authenticate.jsx */
+h3 {
+  margin-bottom: 15px;
+}
+.css-binzgt {
+  margin-top: 0px;
+}
+.exitButton {
+  margin-bottom: 10px;
+  background-color: tomato;
+  padding: 5px 10px 5px 10px;
+  border: none;
+  border-radius: 3px;
+}
+.exitButton:hover {
+  color: white;
+}
+
+.linkButton {
+  margin-bottom: 10px;
+}
+
+#loginSignup {
+  position: absolute;
+  /* top: 50%;
+    left: 50%; */
+  /* transform: translate(-50%, -50%); */
+  background-color: white;
+  /* border: 2px black; */
+  border-radius: 5px;
+  padding: 25px;
+  z-index: 2;
+}
+
+/* Home */
+.home {
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+.homeBanner {
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.homeTitle {
+  font-family: "Nunito", sans-serif;
+  color: white;
+  text-align: center;
+  font-size: 20px;
+}
+
+.homeTitleText {
+  color: tomato;
+  font-size: 50px;
+}
+
+.homeTitleIcon {
+  color: tomato;
+}
+
+/* MapSearch.jsx */
+
+#map {
+  height: 90vh;
+  margin-left: 100px;
+  margin-right: 100px;
+}
+
+.leaflet-container {
+  height: 100%;
+  z-index: 0;
+}
+
+/* Profile.jsx */
+#background {
+  min-height: 100vh;
+  width: auto;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+}
+
+/* ReviewPage.jsx */
+.reviewPageGlobalContainer {
+  margin: 0;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+}
+
+.reviewformContainer {
+  background: rgba(241, 241, 241, 0.6);
+}
+
+.reviewMainContainer {
+  text-decoration: none;
+  font-family: "Roboto", "Helvetica", "Arial", sans-serif;
+  color: rgb(68, 67, 67);
+}
+
+.reviewLabel {
+  margin: 0;
+  text-align: end;
+  justify-content: center;
+  align-items: center;
+  padding-top: 4px;
+}
+
+/* Search.jsx */
+.searchMainContainer {
+  text-decoration: none;
+  font-family: "Roboto", "Helvetica", "Arial", sans-serif;
+  color: rgb(68, 67, 67);
+}
+
+.searchContainer {
+  background: rgba(241, 241, 241, 0.4);
+}
+.searchPageMain {
+  margin: 0;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+}
+
+/* UserProfile.jsx */
+#userProfile {
+  padding: 25px 100px 25px 100px;
+  min-height: 100vh;
+  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-attachment: fixed;
+}
+
+#userProfileTitle {
+  font-family: "Nunito";
+  color: tomato;
+  margin-bottom: 50px;
+}
+h4 {
+  margin-bottom: 20px;
+}
