--- client/public/styles.css
@@ -1,73 +1,0 @@
-#footer {
-  height: 40px;
-  background-color: #333;
-  color: rgba(255, 255, 255, 0.514);
-  font-size: 0.75em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* homeCard.js */
-
-.homeCardContents {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.homeCardRating {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-}
-
-.bikePetFriendlyRating {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-/* homeCards.js */
-
-.homeCardsWrapper {
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeCards {
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  flex-wrap: wrap;
-  margin-bottom: 25px;
-}
-
-.homeCardsUserFeedbackWrapper,
-.homeCardsUserFeedback {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeCardsUserFeedback {
-  margin-bottom: 25px;
-}
-
-.feedbackDesc,
-.feedbackUser {
-  color: white;
-  font-family: "Nunito", sans-serif;
-}
-
-.feedbackDesc {
-  font-style: italic;
-  font-size: 1.1em;
-}
-
-/*  */
