+++ client/src/components/HomeCard.jsx
@@ -1,16 +1,14 @@
 import * as React from "react";
+import { Link } from 'react-router-dom'; 
+
 import Card from "@mui/material/Card";
-import CardActions from "@mui/material/CardActions";
 import CardContent from "@mui/material/CardContent";
 import CardMedia from "@mui/material/CardMedia";
-import Button from "@mui/material/Button";
 import Typography from "@mui/material/Typography";
 import Rating from "@mui/material/Rating";
-import "./homeCard.css";
 import Icon from "@mui/material/Icon";
 import CheckIcon from "@mui/icons-material/Check";
 import ClearIcon from "@mui/icons-material/Clear";
-import { Link } from 'react-router-dom'; 
 
 export default function HomeCard({ landlord }) {
   // console.log("landlord", landlord);
