+++ client/src/components/Navbar.jsx
@@ -1,9 +1,8 @@
 import { Button } from "@mui/material";
 import React, { useState } from "react";
-import "./navbar.css";
 import { NavLink, Link } from "react-router-dom";
-import { Authenticate } from "../../pages/authenticate/Authenticate.jsx";
-import * as actions from "../../actions/actions.js";
+import { Authenticate } from "../pages/Authenticate.jsx";
+import * as actions from "../actions/actions.js";
 
 const mapStateToProps = (state) => ({
   authDisplay: state.display.authDisplay,
