--- client/src/components/navbar/navbar.css
@@ -1,91 +1,0 @@
-/* #loginCover{
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  height: 70vh;
-  width: 100vw;
-  background-color: grey;
-  opacity: 0.3;
-  transform: translate(-50%, -50%);
-  overflow: hidden;
-  box-shadow: 0 0 10px 10px grey;
-
-} */
-
-
-
-/* .linkButton {
-  color:purple;
-  text-decoration: underline;
-}
-
-.linkButton:hover {
-  color:blue;
-} */
-
-#logo {
-  font-size: xx-large;
-  font-family: "Nunito", sans-serif;
-  /* font-style: bold; */
-  color: tomato;
-  font-weight: 700; 
-}
-
-#navBar {
-  height: 50px;
-  width: 100%;
-  display: flex;
-  flex-direction: row;
-  background-color: lightgrey;
-  /* background: none; */
-  align-items: center;
-  /* position: absolute; */
-}
-
-.navBarLeft {
-  flex: 1;
-  display: flex;
-  /* justify-content: center; */
-  padding-left: 25px;
-  align-items: center;
-}
-
-.navBarCenter {
-  flex: 3;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.navBarListItems {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 1.3em;
-}
-
-.navBarListItem {
-  list-style: none;
-  font-family: 'Nunito', sans-serif;
-  padding: 0 20px;
-}
-
-.navBarRight {
-  flex: 1;
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  padding-right: 25px;
-}
-
-
-a:hover {
-  color: tomato;
-} 
-a {
-  font-family: 'Nunito';
-  text-decoration: none;
-  padding-right: 10px;
-  list-style: none;
-  color: #666;
-}
