+++ client/src/components/resultDisplay.jsx
@@ -1,8 +1,8 @@
-import React, { useEffect } from "react";
+import React from "react";
 import Box from "@mui/material/Box";
 
 // import Homecard component
-import HomeCard from "../HomeCard.jsx";
+import HomeCard from "./HomeCard.jsx";
 import { connect } from "react-redux";
 
 const mapStateToProps = (state) => ({
