+++ client/src/index.js
@@ -3,7 +3,8 @@ import { render } from "react-dom";
 import { BrowserRouter } from "react-router-dom";
 import { Provider } from 'react-redux';
 
-import "./public/styles.css"
+import "../public/pageStyles.css"
+import "../public/compStyles.css"
 
 import App from "./App.jsx";
 import store from './store';
