+++ client/src/pages/Home.jsx
@@ -1,12 +1,13 @@
-import "./home.css";
-import HomeCards from "../../components/homeCards/HomeCards.jsx";
-import React, { useEffect, useState } from "react";
-import { Collapse, IconButton, makeStyles } from "@mui/material";
-import { CssBaseline } from "@mui/material";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
+import React, { useEffect } from "react";
 import { Link as Scroll } from "react-scroll";
 import { connect } from "react-redux";
-import * as actions from '../../actions/actions.js';
+
+import { IconButton } from "@mui/material";
+import { CssBaseline } from "@mui/material";
+import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
+
+import HomeCards from "../components/HomeCards.jsx";
+import * as actions from '../actions/actions.js';
 
 const axios = require("axios");
 
@@ -54,4 +55,4 @@ function Home(props) {
 }
 
 
-export default connect(null, mapDispatchtoProps)(Home);
+export default connect(null, mapDispatchToProps)(Home);
