--- client/src/pages/authenticate/authenticate.css
@@ -1,112 +1,0 @@
-h3 {
-  margin-bottom: 15px;
-}
-.css-binzgt {
-  margin-top:  0px;
-}
-.exitButton {
-  margin-bottom: 10px;
-  background-color: tomato;
-  padding: 5px 10px 5px 10px;
-  border: none;
-  border-radius: 3px;
-}
-.exitButton:hover {
-  color: white;
-}
-
-.linkButton{
-  margin-bottom:  10px;
-}
-
-#loginSignup {
-  position: absolute;
-  /* top: 50%;
-  left: 50%; */
-  /* transform: translate(-50%, -50%); */
-  background-color: white;
-  /* border: 2px black; */
-  border-radius: 5px;
-  padding: 25px;
-  z-index: 2;
-}
-/* #loginCover{
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  height: 70vh;
-  width: 100vw;
-  background-color: grey;
-  opacity: 0.3;
-  transform: translate(-50%, -50%);
-  overflow: hidden;
-  box-shadow: 0 0 10px 10px grey;
-
-}
-#loginSignup {
-  opacity: 1;
-  margin:0px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -70%);
-  background-color: white;
-  /* border: 2px black; */
-  /* border-radius: 5px;
-  padding: 25px;
-}
-
-.linkButton {
-  color:purple;
-  text-decoration: underline;
-}
-
-.linkButton:hover {
-  color:blue;
-}
-
-#logo {
-  font-size: xx-large;
-  font-family: "Nunito", sans-serif;
-  font-style: bold;
-  color: tomato;
-  font-weight: 700; 
-}
-
-#navBar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: lightgrey;
-  align-items: center;
-  padding: 15px 50px 15px 50px;
-}
-nav ul {
-  list-style-type: none;  
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  width: 150px;
-  align-self: center;
-  transform: translate(-70%);
-}
-nav ul li {
-  font-size: x-large;
-  margin-left: 20px;
-  margin-right: 50px;
-  list-style: none;
-  margin-left: 50px;
-  color: #666;
-}
-a:hover {
-  color: tomato;
-} 
-a {
-  text-decoration: none;
-   color: black; 
-  font-family: "Roboto","Helvetica","Arial",sans-serif;
-  padding-right: 10px;
-  list-style: none;
-  /* margin-left: 50px; 
-  color: #666;
-} */
