--- client/src/pages/home/home.css
@@ -1,29 +1,0 @@
-.home {
-  min-height: 100vh;
-  background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-}
-
-.homeBanner {
-  height: 100vh;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeTitle {
-  font-family: "Nunito", sans-serif;
-  color: white;
-  text-align: center;
-  font-size: 20px;
-}
-
-.homeTitleText {
-  color: tomato;
-  font-size: 50px;
-}
-
-.homeTitleIcon {
-  color: tomato;
-}
