--- client/src/pages/map/mapSearch.css
@@ -1,11 +1,0 @@
-#map { 
-  height: 90vh;
-  margin-left: 100px;
-  margin-right: 100px;
- }
-
-.leaflet-container {
-  height: 100%;
-  z-index: 0;
-
-}
