--- client/src/pages/review/reviewpage.css
@@ -1,26 +1,0 @@
-.reviewPageGlobalContainer {
-    margin: 0;
-    min-height: 100vh;
-    background-image: url("http://localhost:3000/images/marten-bjork-4WBIyO_2SbU-unsplash.jpg");
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-
-.reviewformContainer{
-    background: rgba(241, 241, 241, 0.6);
-}
-
-.reviewMainContainer {
-    text-decoration: none;
-    font-family: "Roboto","Helvetica","Arial",sans-serif;
-    color: rgb(68, 67, 67);
-}
-
-
-.reviewLabel{
-    margin: 0;
-    text-align: end;
-    justify-content: center;
-    align-items: center;
-    padding-top:4px
-}
