+++ client/src/store.js
@@ -1,9 +1,7 @@
 import { createStore } from 'redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
 import reducers from './reducers/index';
 
 // we are adding composeWithDevTools here to get easy access to the Redux dev tools
 export default store = createStore(
-  reducers,
-  composeWithDevTools()
+  reducers
 );
