+++ webpack.config.js
@@ -43,7 +43,7 @@ module.exports = {
     hot: true,
     magicHtml: true,
     proxy: {
-      'apiKey':'http://localhost:3000',
+      '/apiKey':'http://localhost:3000',
       '/auth': 'http://localhost:3000',
       '/reviews': 'http://localhost:3000',
       '/landlords': 'http://localhost:3000',
